% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootIsotonicRegression.R
\name{bootIsotonicRegression}
\alias{bootIsotonicRegression}
\title{Isotonic Regression Function}
\usage{
bootIsotonicRegression(data, PROBABILITY.GAMMA = 0.5, baselinePava)
}
\arguments{
\item{data}{the same dataframe called by the boot function.}

\item{PROBABILITY.GAMMA}{the target effect probability in the BCD experiment; default = 0.5 and need not be specified.}

\item{baselinePava}{the dataframe prepared by the function preparePava.}
}
\description{
Function of isotonic regression.
}
\examples{
library(ed50)
pavaData <- preparePava(groupS)
bootIsotonicRegression(data = groupS, PROBABILITY.GAMMA = 0.5, baselinePava = pavaData)
}
