% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootIsotonicResample.R
\name{bootIsotonicResample}
\alias{bootIsotonicResample}
\title{The resample function of isotonic regression}
\usage{
bootIsotonicResample(data, mle)
}
\arguments{
\item{data}{Original experiment data.}

\item{mle}{A list of additional arguments to be used by bootIsotonicResample.}
}
\description{
The function is designed as an argument for the boot function of the Canty Bootstrap package.
}
\examples{
library(ed50)
pavaData <- preparePava(groupS)
bootIsotonicResample(data = groupS,
                      mle = list(baselinePava = pavaData,
                                    firstDose = 2.5,
                            PROBABILITY.GAMMA = 0.5))
}
