% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSN.R
\docType{data}
\name{groupSN}
\alias{groupSN}
\title{A Real Experiment Dose Data}
\format{A data of 38 samples and 2 variables:
\describe{
\item{responseSequence}{A value of 0 or 1 indicating the experiment outcome.
0 refers to a failure outcome while 1 refers to a success.}
\item{doseSequence}{The dose given in each experiment.}
}}
\source{
The data is from the article in the references below.
}
\usage{
groupSN
}
\description{
A group of real experiment data based on up-and-down method.
}
\references{
Niu B, Xiao JY, Fang Y, et al. Sevoflurane-induced isoelectric EEG and burst suppression: differential and
antagonistic effect of added nitrous oxide. Anaesthesia 2017; 72: 570-9.
}
\keyword{datasets}
