% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTableOrigin.R
\docType{data}
\name{gTableOrigin}
\alias{gTableOrigin}
\title{G Table}
\format{A data table containing 3 columns:
\describe{
\item{Ratio}{The ratio of dose step and estimate standard error}
\item{G1}{The value of parameter G when the estimate of ED50 falls on a dose level}
\item{G2}{The value of parameter G when the estimate of ED50 falls between two dose levels}
}}
\source{
The table is obtained from Figure 2 in the reference below
}
\usage{
gTableOrigin
}
\description{
A table containing parameter G used in Dixon-Mood method.
}
\references{
Dixon, W. J., & Mood, A. M. (1948). A method for obtaining and analyzing sensitivity 
data. Publications of the American Statistical Association, 43(241), 109-126.
}
\keyword{datasets}
