% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePava.R
\name{preparePava}
\alias{preparePava}
\title{Covert Data Using PAVA Algorithm}
\usage{
preparePava(data)
}
\arguments{
\item{data}{A data frame of dose experiments.}
}
\description{
Covert data using PAVA algorithm, the result is uesd for isotonic regression estimation.
}
\examples{
library(ed50)
preparePava(groupS)
preparePava(groupSN)
}
